refactor: rename style function consistently
