+++ lib/components/DropDown/DropDownOption.tsx
@@ -3,11 +3,11 @@ import clsx from 'clsx';
 import { ComponentProps, FunctionComponent } from 'react';
 import * as React from 'react';
 
+import { textStyles } from '../../styles/typography';
 import { Box } from '../Box/Box';
 import { Icon } from '../Icon/Icon';
 import { Inline } from '../Inline/Inline';
 import { Text } from '../Text/Text';
-import { textStyles } from '../Text/textStyles';
 
 import * as styles from './DropDownOption.css';
