+++ lib/components/Pagination/Bubble.tsx
@@ -7,8 +7,8 @@ import type {
 import * as React from 'react';
 import { ReactNode } from 'react';
 
+import { textStyles } from '../../styles/typography';
 import { Box } from '../Box';
-import { textStyles } from '../Text/textStyles';
 
 import * as styles from './Pagination.css';
