+++ lib/components/Pagination/Pagination.tsx
@@ -8,11 +8,11 @@ import {
 	useMemo,
 } from 'react';
 
+import { textStyles } from '../../styles/typography';
 import { noop } from '../../utils';
 import { Box } from '../Box/Box';
 import { Icon } from '../Icon/Icon';
 import { Inline } from '../Inline/Inline';
-import { textStyles } from '../Text/textStyles';
 
 import { Bubble } from './Bubble';
 import * as styles from './Pagination.css';
