+++ lib/components/Text/index.ts
@@ -1,2 +1,2 @@
 export { Text } from './Text';
-export { styledText as useTextStyles } from '../../styles/typography';
+export { textStyles as useTextStyles } from '../../styles/typography';
