+++ lib/index.ts
@@ -2,5 +2,5 @@ export * from './components';
 export * from './hooks';
 export * from './styles';
 export { overdriveTokens as tokens } from './themes/theme.css';
-export { sprinkles as style } from './styles/sprinkles.css';
+export { sprinkles as css } from './styles/sprinkles.css';
 export { useId, arrayRingLookup } from './utils/index';
